+++ iroh/src/lib.rs
@@ -7,6 +7,7 @@ pub use iroh_net as net;
 pub mod node;
 pub mod rpc_protocol;
 pub mod util;
+pub mod database;
 
 /// Expose metrics module
 #[cfg(feature = "metrics")]
